Make the calculator work
Add event listeners to the operators and the digits separately.

Store the digits and the operators in a storer object that can be passed by reference in the event listener functions.

Nothing happens when an operator is pressed before clicking any number. An operator has a functionality if a number is entered.

The operator function is called by clicking any operator but only if there are two numbers and one operator entered beforehand.

The C button clears all the values of the storer object.

The upperdisplay shows the first number and the operator. If the second number is entered, that is also shown in the upperdisplay. Lower display shows the result and the numbers entered.