Add decimal and change upperDisplay value
Add decimal to the numbers. In order to be able to add the point the number1 and number2 elements are made string and converted to a number with parseFloat when displayed or used as argument in operate function.

Disable decimal button if the numbers include a decimal and enable them again when entering a new number

upperDisplay value should be different when the operator is = sign. For that case, add a conditional before calling the operate function. Call operate function only if the operator is different than = sign.