Change logic of getNumbers and getResult functions
getNumbers: First display the clicked items and then store in number1 or number2 depending on the existence of a stored operator.

Add two more elements to the storer object; result and lastClicked and store the return value of operate in the former and the classList of the last clicked item in the latter.

When an operator is clicked, check if the number entered before has any zeros after the decimal or a decimal not followed by another digit. If so, erase them to simplify the display value. Also check if storer.operator is empty OR two operators that are not '=' are clicked successively. Call getResult if that's not the case.

getResult: First call the operate function and get the result, then modify the display values according to the result and the type of the clicked operator.
If the clicked operator is '=', do not store it so that if a digit is clicked after a calculation, it can be stored in number1.
Add the snarky error message if the user divides by 0 and empty all the elements of the storer object and disable the operator buttons.