Add the functionality of backspace button
First define if display value is representing number1 or number2 by checking the upperDisplay. Then delete the last element of display value and store it in the respective number.

If '=' sign is clicked, the delete button deletes the upper display value and keeps the display value as number1.

Update getNumbers slightly so that the display value won't be 0 when delete is clicked after an operator. Otherwise the display value will be 0 whereas number2 will have a value and that will result the display value to show the numbers with a 0 at the start, such as 0523.