Bind the keyboard
Index.html: Change the id's of C and delete to Escape and Backspace, respectively.
Styles.css: the same as index.html

Script.js: Add eventlistener for keys and the function bindKeyboard. Check all the buttons if their ids match the pressed key. If so, use click() function for that button.