Clean the bugs, except keyboard binding
Backspace bugs: After clicking backspace if the display value was 0, and if the user would click a number, it would display the number after the 0. Correct that within the getNumbers function.
Disable the backspace after the first click when the upper display contains '=', so that it would not delete the numbers in display value one by one. Enable the decimal so that the user can enter a new value. Enable the backspace again in getNumbers and getResult functions.

getNumbers: After an operator if the display value would contain a decimal because number1 has one, clicking the decimal would add another decimal to the display value. Correct that where the display value is stored in number2.

operators addeventlistener: Appoint '0' to number1 when no digit is clicked before an operator, otherwise operate function returns NaN. If the = sign is clicked, don't store it in the storer.operator so that the new clicked values can be stored in number1.