Uncomment keyboard functionality, correct few thng
Change for.each loop to for/of loop so that break can be used when the clickedKey is found.

Erase the upper display value when it has an '=' sign and when backspace is clicked.

Disable the text field so that the user can't click on it or enter an input. Style the disabled text field on css.