Debug keyboard binding and clean a few things
When clicked a button, and pressed enter on keyboard the clicked button stayed focused. Add e.preventDefault() to prevent that behaviour.

Convert all == to ===.

Add displayInput function to call in getNumbers function.

Store '=' sign in store.operator and change the conditionals accordingly to make the code more readable.

Math.round the storer.result to prevent floating numbers but it is still not solved for large floating numbers.

Add cleanValues function to clean the unnecessary zeros and the decimal and call it in operators eventlistener